requirements.txt
# Web Framework
fastapi~=0.115.0
uvicorn~=0.30.0

# Cliente HTTP para comunicación entre microservicios
httpx~=0.28.0

# Criptografía (Algoritmos modernos: AES, DES, 3DES, ChaCha20, RSA, ECC, etc.)
pycryptodome~=3.20.0

# Base de Datos Relacional (PostgreSQL)
asyncpg~=0.30.0
psycopg2-binary  # Sin versión fija para evitar problemas de compilación en Windows

# Base de Datos No Relacional (MongoDB)
pymongo~=4.10.0
motor~=3.6.0

# Utilidades
pydantic~=2.0.0
python-dotenv~=1.0.0
pytz~=2024.0

# Testing (opcional - descomentar si se necesita)
# pytest~=8.0.0
# pytest-asyncio~=0.25.0
# pytest-httpx~=0.35.0

python -m pip install pycryptodome